Freeman Barrel position fix
The second installment of the Half-Life series, Half-Life 2, was released on November 16th, 2004 by Valve Corporation. It was a tech hallmark back then and was considered the "Game of the Decade" by multiple gaming outlets at the time. Its legacy still carries on today as one of the most-influential FPS games.