Sound Optimizations for 1911.
This will be the first of my efforts to cut down the file sizes of our lead sound producers Rzen1th's sound's. The goal is to make the changes not super noticeable and still make gunplay fun and punchy sounding without sacrificing too much quality. Feel free to tell me if you notice anything that's odd sounding.

Notes: The tail sounds made up the bulk of the file size. If I made the sound quality too low you would get this very wumpy and wooshy sound. So I just settled for 8 quality instead of 10. Bit iffy on the change but it sounds alright.