Remade Deagle Fire's
tldr: tried making it sound like the original but nerfed the power a bit.

Taking a break from being online for a bit I decided to remake the deagle. Lead SFX Guy Rzen1th wanted me to do this anyways so here we go. I wanted to make it sound like the original but nerf it a little. The original in my opinion sounded way too strong and almost rifle like. I wanted to bring more of that handgun sound into it and this was the result.

Might change it later. The mono layers throw the sound off a wee bit.